Refactoring wrong location of frontend code
Unnecesary service class deleted

Cleaning code

Cleaning code
