+++ src/app/core/layout/language-changer/language-changer.spec.ts
@@ -11,11 +11,6 @@ import { LanguagesModel } from './language-changer.interface';
 
 describe('LanguageChangerComponent', () => {
 
-/*   test('',() => {
-    expect(true).toBe(true);
-  }); */
-
-
   let component: LanguageChangerComponent;
   let fixture: ComponentFixture<LanguageChangerComponent>;
