+++ src/app/features/home/page/home.component.spec.ts
@@ -8,12 +8,6 @@ import { HomeComponent } from './home.component';
 
 describe('HomeComponent', () => {
 
-
-/*   test('',() => {
-    expect(true).toBe(true);
-  }); */
-  
-
   let component: HomeComponent;
   let fixture: ComponentFixture<HomeComponent>;
