--- src/app/features/my-environment/pages/my-environment-page/environment-models.ts
@@ -1,5 +1,0 @@
-
-export interface Icons {
-  viewbox:string;
-  icon:string
-}
