+++ src/app/features/my-environment/pages/my-environment-page/my-environment-page.component.spec.ts
@@ -17,10 +17,6 @@ const routes: Routes = [
 
 describe('MyEnvironmentPageComponent', () => {
 
- /*  test('',() => {
-    expect(true).toBe(true);
-  });
- */
   const { scroll } = window;
 
   beforeAll(() => {
