+++ src/app/features/my-environment/pages/my-environment-search/my-environment-search.component.spec.ts
@@ -21,10 +21,6 @@ const routes: Routes = [
 
 describe('MyEnvironmentSearchComponent', () => {
 
-/*   test('',() => {
-    expect(true).toBe(true);
-  }); */
-
   let component: MyEnvironmentSearchComponent;
   let fixture: ComponentFixture<MyEnvironmentSearchComponent>;
   let myEnvSrv: MyEnvironmentService
