+++ src/app/features/users/components/register/register.component.spec.ts
@@ -4,11 +4,6 @@ import { RegisterComponent } from './register.component';
 
 describe('RegisterComponent', () => {
 
-  test('',() => {
-    expect(true).toBe(true);
-  });
-
-
   let component: RegisterComponent;
   let fixture: ComponentFixture<RegisterComponent>;
